requirements.txt
# A2A Protocol & Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Queue System
redis>=5.0.0

# RAG & Vector Store
chromadb>=0.4.22
sentence-transformers>=2.2.2

# LLM Providers
google-generativeai>=0.8.5
anthropic>=0.40.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0

# Logging
structlog>=24.1.0
